refactor(indents): migrate components to feature directory and extract IndentTable (Phase 6c)
- Move 'IndentRow' from shared resources to 'features/indents/components'.
- Extract 'IndentTable' component to handle table layout and pagination logic.
- Update indents list page to use new high-level feature components.
