+++ app/users/dashboard/resources/indents/page.tsx
@@ -2,28 +2,10 @@
 
 import { useState, useMemo } from 'react';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
 import { routes } from '@/nav';
-import { Card, CardContent, CardHeader } from '@/components/shadcn/card';
+import { Card } from '@/components/shadcn/card';
 import { Button } from '@/components/shadcn/button';
-import { Badge } from '@/components/shadcn/badge';
-import { Input } from '@/components/shadcn/input';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/shadcn/select';
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/shadcn/table';
-import { Pagination, PageHeader } from '@/components/common';
+import { PageHeader } from '@/components/common';
 import {
   Plus,
   Loader2,
@@ -31,30 +13,19 @@ import {
   Clock,
   TruckIcon,
   Package,
-  FolderOpen,
-  Search,
 } from 'lucide-react';
 import {
   Empty,
-  EmptyMedia,
   EmptyErrorMedia,
   EmptyHeader,
   EmptyTitle,
   EmptyDescription,
 } from '@/components/shadcn/empty';
-import { format } from 'date-fns';
 import { useIndentsPaginated } from '@/hooks/indents';
-import { IndentRow } from '@/features/resources/components';
-import {
-  IndentStatus,
-  indentStatusLabels,
-  indentStatusBadgeColors,
-} from '@/types/indents';
-
-const PAGE_SIZE_OPTIONS = [5, 10, 20, 50, 100];
+import { IndentTable } from '@/features/indents/components';
+import { IndentStatus } from '@/types/indents';
 
 export default function IndentsPage() {
-  const router = useRouter();
   const {
     data: indents = [],
     isLoading,
@@ -66,22 +37,6 @@ export default function IndentsPage() {
   const [projectFilter, setProjectFilter] = useState('all');
   const [currentPage, setCurrentPage] = useState(1);
   const [itemsPerPage, setItemsPerPage] = useState(10);
-  const [prevFilters, setPrevFilters] = useState({
-    searchQuery,
-    statusFilter,
-    projectFilter,
-    itemsPerPage,
-  });
-
-  if (
-    prevFilters.searchQuery !== searchQuery ||
-    prevFilters.statusFilter !== statusFilter ||
-    prevFilters.projectFilter !== projectFilter ||
-    prevFilters.itemsPerPage !== itemsPerPage
-  ) {
-    setPrevFilters({ searchQuery, statusFilter, projectFilter, itemsPerPage });
-    setCurrentPage(1);
-  }
 
   const filtered = useMemo(() => {
     const q = searchQuery.toLowerCase();
@@ -108,8 +63,7 @@ export default function IndentsPage() {
 
   const totalPages = Math.ceil(filtered.length / itemsPerPage);
   const startIndex = (currentPage - 1) * itemsPerPage;
-  const endIndex = Math.min(startIndex + itemsPerPage, filtered.length);
-  const paginated = filtered.slice(startIndex, endIndex);
+  const paginated = filtered.slice(startIndex, startIndex + itemsPerPage);
 
   const stats = {
     total: indents.length,
@@ -160,7 +114,6 @@ export default function IndentsPage() {
         }
       />
 
-      {/* Stats */}
       <Card className="gap-0 p-6">
         <div className="sm:divide-border grid grid-cols-2 gap-3 sm:grid-cols-4 sm:gap-0 sm:divide-x">
           <div className="flex flex-col gap-1 rounded-lg p-3 sm:rounded-none sm:pr-6">
@@ -222,148 +175,36 @@ export default function IndentsPage() {
         </div>
       </Card>
 
-      {/* Table Card */}
-      <Card>
-        <CardHeader className="flex flex-row items-center gap-3 border-b px-4 py-1">
-          <div className="relative w-full max-w-xs">
-            <Search className="absolute top-1/2 left-3 size-3.5 -translate-y-1/2 text-zinc-400" />
-            <Input
-              value={searchQuery}
-              onChange={(e) => {
-                setSearchQuery(e.target.value);
-                setCurrentPage(1);
-              }}
-              placeholder="Search by indent number, project or creator…"
-              className="h-8 pl-8 text-sm"
-            />
-          </div>
-          <Select
-            value={statusFilter}
-            onValueChange={(v) => {
-              setStatusFilter(v);
-              setCurrentPage(1);
-            }}
-          >
-            <SelectTrigger className="h-8 w-36 text-xs">
-              <SelectValue placeholder="Status" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="all">All Statuses</SelectItem>
-              {Object.values(IndentStatus).map((s) => (
-                <SelectItem key={s} value={s}>
-                  {indentStatusLabels[s]}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-          <Select
-            value={projectFilter}
-            onValueChange={(v) => {
-              setProjectFilter(v);
-              setCurrentPage(1);
-            }}
-          >
-            <SelectTrigger className="h-8 w-36 text-xs">
-              <SelectValue placeholder="Project" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="all">All Projects</SelectItem>
-              {projectOptions.map((p) => (
-                <SelectItem key={p} value={p}>
-                  {p}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-          <div className="ml-auto flex items-center gap-2 border-l pl-3">
-            <span className="text-xs whitespace-nowrap text-zinc-500">
-              Rows per page
-            </span>
-            <Select
-              value={String(itemsPerPage)}
-              onValueChange={(v) => {
-                setItemsPerPage(Number(v));
-                setCurrentPage(1);
-              }}
-            >
-              <SelectTrigger className="h-8 w-16 text-xs">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                {PAGE_SIZE_OPTIONS.map((n) => (
-                  <SelectItem key={n} value={String(n)}>
-                    {n}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-        </CardHeader>
-
-        {paginated.length > 0 ? (
-          <>
-            <CardContent className="p-0">
-              <Table>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className="pl-6">Indent #</TableHead>
-                    <TableHead>Status</TableHead>
-                    <TableHead>Project</TableHead>
-                    <TableHead>Items</TableHead>
-                    <TableHead>Expected On</TableHead>
-                    <TableHead className="pr-6">Created By</TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {paginated.map((indent) => (
-                    <IndentRow
-                      key={indent.id}
-                      indent={indent}
-                      onClick={() =>
-                        router.push(
-                          routes.resources.indents.detail(indent.id).href
-                        )
-                      }
-                    />
-                  ))}
-                </TableBody>
-              </Table>
-            </CardContent>
-            <div className="flex items-center justify-between border-t px-4 py-2">
-              <span className="text-sm text-zinc-500">
-                {startIndex + 1}–{endIndex} of {filtered.length} indent
-                {filtered.length === 1 ? '' : 's'}
-              </span>
-              <Pagination
-                currentPage={currentPage}
-                totalPages={totalPages}
-                onPageChange={setCurrentPage}
-              />
-            </div>
-          </>
-        ) : (
-          <CardContent>
-            <Empty variant="default">
-              <EmptyMedia variant="icon">
-                <ClipboardList className="size-6" />
-              </EmptyMedia>
-              <EmptyHeader>
-                <EmptyTitle>No indents found</EmptyTitle>
-                <EmptyDescription>
-                  {hasActiveFilters
-                    ? 'No indents match your filters. Try adjusting your search.'
-                    : 'Create your first indent to get started.'}
-                </EmptyDescription>
-              </EmptyHeader>
-              {!hasActiveFilters && (
-                <Button asChild>
-                  <Link href={routes.resources.indents.new}>New Indent</Link>
-                </Button>
-              )}
-            </Empty>
-          </CardContent>
-        )}
-      </Card>
+      <IndentTable
+        paginated={paginated}
+        filteredCount={filtered.length}
+        startIndex={startIndex}
+        itemsPerPage={itemsPerPage}
+        onItemsPerPageChange={(n) => {
+          setItemsPerPage(n);
+          setCurrentPage(1);
+        }}
+        currentPage={currentPage}
+        totalPages={totalPages}
+        onPageChange={setCurrentPage}
+        hasActiveFilters={hasActiveFilters}
+        searchValue={searchQuery}
+        onSearchChange={(v) => {
+          setSearchQuery(v);
+          setCurrentPage(1);
+        }}
+        statusFilter={statusFilter}
+        onStatusChange={(v) => {
+          setStatusFilter(v);
+          setCurrentPage(1);
+        }}
+        projectFilter={projectFilter}
+        onProjectChange={(v) => {
+          setProjectFilter(v);
+          setCurrentPage(1);
+        }}
+        projectOptions={projectOptions}
+      />
     </div>
   );
 }
