+++ features/indents/components/index.ts
@@ -2,3 +2,5 @@ export * from './indent-alert-dialogs';
 export * from './indent-items-card';
 export * from './indent-info-card';
 export * from './indent-remarks-card';
+export { IndentTable } from './indent-table';
+export { IndentRow } from './indent-row';
